requirements-runtime.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
python-multipart>=0.0.9
pyyaml>=6.0
python-dotenv>=1.0
structlog>=24.0
orjson>=3.9
pypdf>=4.0
tiktoken>=0.6
numpy>=1.26
scipy>=1.11
chromadb>=0.5
openai>=1.0
tenacity>=8.2
rank-bm25>=0.2
sentence-transformers>=2.7
prometheus-client>=0.20
httpx>=0.27
